# api/openapi.yaml
openapi: 3.0.3
info:
  title: Darkory
  version: 0.1.0
  description: |
    The canonical contract between Members and a Darkory Install. The CLI, the MCP server and the
    web app use only these operations; humans and agents call the same ones.

    **Versioning.** Paths are versioned (`/v1`). Changes within `/v1` are additive only.

    **Credentials.** Every operation except `getHealth`, `showLoginLink`, `redeemLoginLink` and
    `requestEmailSignIn` needs a credential: a bearer token (`dk_…`) together with a
    `Darkory-Session` header naming the Session making the request, or the browser cookie
    `darkory_session`. A bearer request without `Darkory-Session` is refused with
    `session_required`. A request with no credential is never a Member. A browser Session ends
    after a time unused and after a time in all (Install settings); a deactivated Member's
    credentials are all refused.

    **References.** Path parameters and request-body references take an id, or the natural key
    where one exists: a Task or Feature display key (`WEB-42`), a Team key (`WEB`), a Skill name
    (`qa`), a Member name. Tokens, Evidence and Sessions are referenced by id only. Responses
    always carry ids, plus display keys where they exist.

    **Retries.** Every write accepts an `Idempotency-Key` header. The first response is kept for
    24 hours per Member and key; a retry with the same key returns it unchanged. Reusing a key
...

...

...
    omitted.
  license:
    name: Apache-2.0
...
servers:
...
security:
  - bearerAuth: []
    sessionHeader: []
  - cookieAuth: []
...
paths:
...
  /v1/features:
    post:
      operationId: fileFeature
      tags: [features]
      summary: File a Feature
      description: |
        Files the Feature at the bottom of its Team's Rank, and files its Break down Task needing
        the `breakdown` Skill in the same write. The owner defaults to the caller. A Feature filed
        by a Retrospective names it in `from_retrospective`. Errors: `forbidden` (not in the Team).
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/FileFeatureBody"}
      responses:
        "201":
          description: The Feature with its Break down Task.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FeatureDetail"}
        default: {$ref: "#/components/responses/Error"}
    get:
      operationId: listFeatures
      tags: [features]
      summary: List Features
      description: Ordered by Team, then Rank.
      parameters:
        - name: team
          in: query
          schema: {type: string}
        - name: state
          in: query
          schema: {$ref: "#/components/schemas/FeatureState"}
        - name: owner
          in: query
          schema: {type: string}
        - {$ref: "#/components/parameters/Limit"}
        - {$ref: "#/components/parameters/Cursor"}
      responses:
        "200":
          description: The Features.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FeatureList"}
        default: {$ref: "#/components/responses/Error"}
  /v1/features/{feature}:
    get:
      operationId: getFeature
      tags: [features]
      summary: Get a Feature with its Tasks and Evidence
      parameters:
        - {$ref: "#/components/parameters/FeatureRef"}
      responses:
        "200":
          description: The Feature.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FeatureDetail"}
        default: {$ref: "#/components/responses/Error"}
...
  /v1/tasks:
    post:
      operationId: fileTask
      tags: [tasks]
      summary: File a Task
      description: |
        A Task needs a Skill or is aimed at a Member by name, not both. Naming `blocks` files a
        question or Escalation: the new Task joins the Feature of the Task it blocks (which must
        then be the `feature` given, or `feature` may be left out) and blocks it in the same write,
        even when that Feature has ended; the asker keeps their Claim. Blocking a Task needs its
        Claim when it is held, else its Feature's ownership or membership of its Team. Errors:
        `ended` (the Feature has ended and the Task blocks nothing, or the blocked Task has ended),
        `not_holder`, `forbidden`, `cycle`.
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/FileTaskBody"}
      responses:
        "201":
          description: The new Task.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TaskDetail"}
        default: {$ref: "#/components/responses/Error"}
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List Tasks
      description: Ordered by Feature Rank, then by how long each Task has waited.
      parameters:
        - name: feature
          in: query
          schema: {type: string}
        - name: team
          in: query
          schema: {type: string}
        - name: state
          in: query
          schema: {$ref: "#/components/schemas/TaskState"}
        - name: skill
          in: query
          description: Only Tasks that need this Skill now.
          schema: {type: string}
        - name: aimed_at
          in: query
          description: Only Tasks aimed at this Member.
          schema: {type: string}
        - name: holder
          in: query
          description: Only Tasks this Member holds a live Claim on.
          schema: {type: string}
        - {$ref: "#/components/parameters/Limit"}
        - {$ref: "#/components/parameters/Cursor"}
      responses:
        "200":
          description: The Tasks.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TaskList"}
        default: {$ref: "#/components/responses/Error"}
  /v1/tasks/takeable:
    get:
      operationId: listTakeableTasks
      tags: [tasks]
      summary: List the Tasks the caller can take now, in the order `next` would offer them
      parameters:
        - {$ref: "#/components/parameters/Limit"}
      responses:
        "200":
          description: The takeable Tasks.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TaskList"}
        default: {$ref: "#/components/responses/Error"}
  /v1/tasks/next:
    post:
      operationId: nextTask
      tags: [tasks]
      summary: Wait for a takeable Task and claim it
      description: |
        Claims the first Task takeable by the caller, in Rank order across the caller's Teams
        (a tie goes to the Task that has waited longest; within a Feature, Tasks that block
        another come first). When none is takeable, holds the request open for up to
        `wait_seconds` and claims one as soon as it becomes takeable; replies 204 when the wait
        ends with nothing claimed. The Claim takes `heartbeat_timeout_seconds`, or else the
        token's default. A Member may have a limited number of `next` calls waiting at once
        (an Install setting, 16 by default); one more is refused with `too_many_requests`.
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/NextTaskBody"}
      responses:
        "200":
          description: The claimed Task.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TaskDetail"}
        "204":
          description: Nothing became takeable before the wait ended.
        default: {$ref: "#/components/responses/Error"}
  /v1/tasks/{task}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a Task with its Claims, Notes, Evidence, blockers and Observations
      parameters:
        - {$ref: "#/components/parameters/TaskRef"}
      responses:
        "200":
          description: The Task.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TaskDetail"}
        default: {$ref: "#/components/responses/Error"}
  /v1/tasks/{task}/claim:
    post:
      operationId: claimTask
      tags: [tasks]
      summary: Claim a Task
      description: |
        One conditional write: it succeeds only when the Task is takeable by the caller. The
        Claim takes `heartbeat_timeout_seconds`, or else the token's default; with a timeout it is
        bound to the calling Session, without one to the Member. Errors: `already_claimed`
        (someone holds it; stop rather than retry), `not_takeable`.
      parameters:
        - {$ref: "#/components/parameters/TaskRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ClaimTaskBody"}
      responses:
        "200":
          description: The claimed Task.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TaskDetail"}
        default: {$ref: "#/components/responses/Error"}
  /v1/tasks/{task}/heartbeat:
    post:
      operationId: heartbeat
      tags: [tasks]
      summary: Tell Darkory the caller is still working a Task
      description: |
        Extends a Claim that carries a heartbeat timeout. A Heartbeat that arrives after the
        timeout is refused and the reply says `lapsed`; the Member may claim again if the Task is
        still takeable. Writes no Activity. Repeating a Heartbeat is harmless, so its
        `Idempotency-Key` is accepted but not stored: a retry reports the Claim as it is now.
        Errors: `not_holder` (the caller never held this Claim).
      parameters:
        - {$ref: "#/components/parameters/TaskRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "200":
          description: What became of the caller's Claim.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/HeartbeatReply"}
        default: {$ref: "#/components/responses/Error"}
  /v1/tasks/{task}/release:
    post:
      operationId: releaseTask
      tags: [tasks]
      summary: Give up the caller's Claim, leaving the Task needing the same Skill
      description: "Errors: `not_holder`."
      parameters:
        - {$ref: "#/components/parameters/TaskRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ReleaseTaskBody"}
      responses:
        "200":
          description: The Task.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Task"}
        default: {$ref: "#/components/responses/Error"}
...
  /v1/tasks/{task}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Complete a Task the caller holds
      description: |
        Ends the Task done. Completing a Task that needs `skill-review` and carries a pending
        proposal publishes it as the Skill's next version; completing a Retrospective marks its
        Feature's unreviewed Observations reviewed by it. Errors: `not_holder`, `proposal_stale`
        (the version the proposal was written against is no longer current; nothing changes, and
        the Task is handed back to `retro`), `forbidden` (the caller wrote the proposal).
      parameters:
        - {$ref: "#/components/parameters/TaskRef"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CompleteTaskBody"}
      responses:
        "200":
          description: The Task.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Task"}
        default: {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A Member token, `dk_…`. Sent together with `Darkory-Session`.
    sessionHeader:
      type: apiKey
      in: header
      name: Darkory-Session
      description: The id the running copy chose for its Session. Required with a bearer token.
    cookieAuth:
      type: apiKey
      in: cookie
      name: darkory_session
      description: The browser Session, set by redeeming a login link.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        A key unique to this write. A retry with the same key returns the first response. It is 1
        to 255 printable ASCII characters, without spaces; any other is refused with `invalid`.
      schema: {type: string, minLength: 1, maxLength: 255, pattern: "^[!-~]+$"}
...
    FeatureRef:
      name: feature
      in: path
      required: true
      description: Feature id or display key, such as `WEB-1`.
      schema: {type: string}
    TaskRef:
      name: task
      in: path
      required: true
      description: Task id or display key, such as `WEB-42`.
      schema: {type: string}
...
    Limit:
      name: limit
      in: query
      description: At most this many items. Defaults to 100.
      schema: {type: integer, minimum: 1, maximum: 500}
    Cursor:
      name: cursor
      in: query
      description: The `next_cursor` of the previous page.
      schema: {type: string}
  responses:
    Error:
      description: The request failed.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: {$ref: "#/components/schemas/ErrorCode"}
        message:
          type: string
          description: For people; may change between releases.
        details:
          type: object
          additionalProperties: true
          description: Extra facts about the failure, by code.
    ErrorCode:
      type: string
      description: |
        Stable reason for a failure, with its HTTP status:
        `invalid` 400 · `unauthenticated` 401 · `session_required` 401 · `forbidden` 403 ·
        `not_found` 404 · `conflict` 409 · `already_claimed` 409 · `not_takeable` 409 ·
        `not_holder` 409 · `ended` 409 · `tasks_open` 409 · `cycle` 409 · `proposal_stale` 409 ·
        `too_large` 413 · `idempotency_key_reused` 422 · `too_many_requests` 429 · `internal` 500 ·
        `not_implemented` 501.
      enum:
        - invalid
        - unauthenticated
        - session_required
        - forbidden
        - not_found
        - conflict
        - already_claimed
        - not_takeable
        - not_holder
        - ended
        - tasks_open
        - cycle
        - proposal_stale
        - too_large
        - idempotency_key_reused
        - too_many_requests
        - internal
        - not_implemented

    # ------------------------------------------------------------ install
...
    SkillProposal:
      type: object
      required: [id, skill_id, task_id, based_on_version, body, author_id, state, created_at]
      properties:
        id: {type: string}
        skill_id: {type: string}
        task_id: {type: string}
        based_on_version:
          type: integer
          format: int64
        body: {type: string}
        author_id: {type: string}
        state: {$ref: "#/components/schemas/ProposalState"}
        published_version:
          type: integer
          format: int64
          description: The version publishing it made. Absent unless `state` is `published`.
        created_at: {type: string, format: date-time}
        decided_at:
          type: string
          format: date-time
          description: When it was published or superseded.
    ProposalState:
      type: string
      description: |
        `pending`: waiting for review. `published`: a review published it. `superseded`: it will
        not be published, because a newer proposal replaced it on its Task or its Task ended
        without publishing it.
      enum: [pending, published, superseded]
...
    Feature:
      type: object
      required: [id, key, team_id, title, description, owner_id, state, rank, filed_by, created_at, task_counts]
      properties:
        id: {type: string}
        key:
          type: string
          description: Display key, such as `WEB-1`.
        team_id: {type: string}
        title: {type: string}
        description: {type: string}
        owner_id: {type: string}
        state: {$ref: "#/components/schemas/FeatureState"}
        rank:
          type: integer
          format: int64
          description: Position in the Team's Rank, 1 first. An ended Feature keeps its place.
        from_retrospective_task_id:
          type: string
          description: The Retrospective that filed this Feature.
        filed_by: {type: string}
        created_at: {type: string, format: date-time}
        ended_at: {type: string, format: date-time}
        task_counts: {$ref: "#/components/schemas/TaskCounts"}
    TaskCounts:
      type: object
      description: How many of the Feature's Tasks are in each state.
      required: [open, claimed, done, dropped]
      properties:
        open:
          type: integer
          description: Open Tasks, claimed or not.
        claimed:
          type: integer
          description: Open Tasks with a live Claim; these are also counted in `open`.
        done: {type: integer}
        dropped: {type: integer}
    FeatureState:
      type: string
      enum: [open, shipped, dropped]
    FeatureDetail:
      type: object
      required: [feature, tasks, evidence]
      properties:
        feature: {$ref: "#/components/schemas/Feature"}
        tasks:
          type: array
          items: {$ref: "#/components/schemas/Task"}
        evidence:
          type: array
          items: {$ref: "#/components/schemas/Evidence"}
    FeatureList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/Feature"}
        next_cursor:
          type: string
          description: Pass as `cursor` for the next page. Absent on the last page.
    FileFeatureBody:
      type: object
      required: [team, title]
      properties:
        team:
          type: string
          description: Team id or key.
        title: {type: string, minLength: 1, maxLength: 200}
        description: {type: string}
        owner:
          type: string
          description: Member id or name. Defaults to the caller.
        from_retrospective:
          type: string
          description: Id or display key of the Retrospective Task filing this Feature.
...
    Task:
      type: object
      required: [id, key, feature_id, kind, title, description, state, blocked, filed_by, waiting_since, created_at]
      properties:
        id: {type: string}
        key:
          type: string
          description: Display key, such as `WEB-42`.
        feature_id: {type: string}
        kind: {$ref: "#/components/schemas/TaskKind"}
        title: {type: string}
        description: {type: string}
        state: {$ref: "#/components/schemas/TaskState"}
        skill_id:
          type: string
          description: The Skill the Task needs now. Absent when it is aimed at a Member.
        aimed_at_id:
          type: string
          description: The Member the Task is aimed at by name.
        claim:
          $ref: "#/components/schemas/Claim"
        blocked:
          type: boolean
          description: True while any Task blocking this one is open.
        open_blockers:
          type: array
          description: The open Tasks blocking this one. Absent when none is open.
          items: {$ref: "#/components/schemas/TaskBrief"}
        filed_by: {type: string}
        waiting_since:
          type: string
          format: date-time
          description: When the Task was filed or last handed over.
        created_at: {type: string, format: date-time}
        ended_at: {type: string, format: date-time}
    TaskBrief:
      type: object
      description: A Task named by its id and display key.
      required: [id, key]
      properties:
        id: {type: string}
        key:
          type: string
          description: Display key, such as `WEB-42`.
    TaskState:
      type: string
      description: Claimed and lapsed are not states; they follow from the Task's Claim.
      enum: [open, done, dropped]
    TaskKind:
      type: string
      description: "`breakdown` and `retrospective` Tasks are filed by Darkory."
      enum: [work, breakdown, retrospective]
    Claim:
      type: object
      required: [id, task_id, holder_id, session_id, started_at]
      properties:
        id: {type: string}
        task_id: {type: string}
        holder_id: {type: string}
        session_id:
          type: string
          description: The Session that made the Claim.
        skill_id:
          type: string
          description: The Skill the Claim was made under. Absent for a Task aimed at a Member.
        skill_version:
          type: integer
          format: int64
          description: The version of the Skill the Claim was worked under.
        model_label:
          type: string
          description: The AI model the holder said it used. Stored, never interpreted.
        heartbeat_timeout_seconds:
          type: integer
          description: Absent when the Claim never lapses on its own.
        started_at: {type: string, format: date-time}
        expires_at:
          type: string
          format: date-time
          description: When the Claim lapses unless a Heartbeat arrives.
        ended_at: {type: string, format: date-time}
        how_ended: {$ref: "#/components/schemas/ClaimEnd"}
    ClaimEnd:
      type: string
      enum: [released, handed_over, completed, lapsed, taken_back, dropped, token_revoked, session_closed, member_deactivated]
    TaskDetail:
      type: object
      required: [task, feature, claims, notes, evidence, blockers, blocking, observations]
      description: The Task with its record. `proposal` is the latest Skill proposal written on it, when any.
      properties:
        task: {$ref: "#/components/schemas/Task"}
        feature: {$ref: "#/components/schemas/Feature"}
        claims:
          type: array
          description: Every Claim on the Task, oldest first.
          items: {$ref: "#/components/schemas/Claim"}
        notes:
          type: array
          description: The running log, oldest first.
          items: {$ref: "#/components/schemas/Note"}
        evidence:
          type: array
          items: {$ref: "#/components/schemas/Evidence"}
        blockers:
          type: array
          description: The Tasks blocking this one.
          items: {$ref: "#/components/schemas/Task"}
        blocking:
          type: array
          description: The Tasks this one blocks.
          items: {$ref: "#/components/schemas/Task"}
        observations:
          type: array
          items: {$ref: "#/components/schemas/Observation"}
        proposal:
          $ref: "#/components/schemas/SkillProposal"
    TaskList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/Task"}
        next_cursor:
          type: string
          description: Pass as `cursor` for the next page. Absent on the last page.
    FileTaskBody:
      type: object
      required: [title]
      properties:
        feature:
          type: string
          description: Feature id or display key. May be left out when `blocks` is given.
        title: {type: string, minLength: 1, maxLength: 200}
        description: {type: string}
        skill:
          type: string
          description: Skill id or name the Task needs. Give this or `aimed_at`.
        aimed_at:
          type: string
          description: Member id or name the Task is aimed at. Give this or `skill`.
        blocks:
          type: string
          description: Id or display key of a Task the new one blocks (a question or Escalation).
    ClaimTaskBody:
      type: object
      properties:
        heartbeat_timeout_seconds:
          type: integer
          minimum: 0
          maximum: 86400
          description: Seconds without a Heartbeat before the Claim lapses. 0 for none. Defaults to the token's default.
        model_label:
          type: string
          maxLength: 200
    NextTaskBody:
      type: object
      properties:
        wait_seconds:
          type: integer
          minimum: 0
          maximum: 60
          description: How long to wait for a takeable Task. Defaults to 30; 0 returns at once.
        heartbeat_timeout_seconds:
          type: integer
          minimum: 0
          maximum: 86400
          description: Seconds without a Heartbeat before the Claim lapses. 0 for none. Defaults to the token's default.
        model_label:
          type: string
          maxLength: 200
    HeartbeatReply:
      type: object
      required: [status]
      properties:
        status: {$ref: "#/components/schemas/HeartbeatStatus"}
        expires_at:
          type: string
          format: date-time
          description: The Claim's new expiry, when `status` is `ok`.
    HeartbeatStatus:
      type: string
      description: |
        `ok`: the Claim is extended. `lapsed`: it ran out before this Heartbeat. `taken_back`:
        someone on the Reporting line or the Feature owner took it back. `ended`: it ended some
        other way (released, handed over, completed, dropped, revoked).
      enum: [ok, lapsed, taken_back, ended]
    ReleaseTaskBody:
      type: object
      properties:
        note:
          type: string
          description: Added to the Task's Notes in the same write.
...
    CompleteTaskBody:
      type: object
      properties:
        note:
          type: string
          description: Added to the Task's Notes in the same write.
...
    Note:
      type: object
      required: [id, task_id, author_id, body, created_at]
      properties:
        id: {type: string}
        task_id: {type: string}
        author_id: {type: string}
        skill_id:
          type: string
          description: The Skill the author was working under.
        body: {type: string}
        created_at: {type: string, format: date-time}
...
    Observation:
      type: object
      required: [id, task_id, feature_id, author_id, outcome, body, created_at]
      properties:
        id: {type: string}
        task_id: {type: string}
        feature_id: {type: string}
        author_id: {type: string}
        skill_id:
          type: string
          description: The Skill the author was working under.
        outcome: {$ref: "#/components/schemas/ObservationOutcome"}
        body: {type: string}
        created_at: {type: string, format: date-time}
        reviewed_by_task_id:
          type: string
          description: The Retrospective that reviewed it. Absent until reviewed.
        reviewed_at: {type: string, format: date-time}
    ObservationOutcome:
      type: string
      enum: [worked, didnt_work]
...
    Evidence:
      type: object
      required: [id, feature_id, filename, content_type, size, sha256, attached_by, created_at]
      properties:
        id: {type: string}
        feature_id: {type: string}
        task_id:
          type: string
          description: Absent when attached to the Feature itself.
        filename: {type: string}
        content_type: {type: string}
        size:
          type: integer
          format: int64
        sha256: {type: string}
        attached_by: {type: string}
        created_at: {type: string, format: date-time}

    # ------------------------------------------------------------ activity
...